requirements.txt
Django==2.2.3
django-jet==1.0.8
django-ordered-model==3.1.1
django-taggit==1.1.0
psycopg2-binary==2.8.3
pytz==2019.1
sqlparse==0.3.0
